pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10,<3.13"
dependencies = [
    "apache-airflow[celery]==2.9.0",
    "azure-storage-file-datalake>=12.23.0",
    "deltalake>=1.4.2",
    "duckdb>=1.5.0",
    "matplotlib>=3.10.8",
    "pandas>=2.0,<3.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
[tool.setuptools]
package-dir = {"" = "src"}
